+++ src/Transformer/String/AsciiLettersOnly/AsciiLettersOnlyTransformer.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHP\Transformer\String\AsciiLettersOnly;
+
+use ExtendsSoftware\ExaPHP\Transformer\TransformerInterface;
+
+use function preg_replace;
+
+class AsciiLettersOnlyTransformer implements TransformerInterface
+{
+    /**
+     * @inheritDoc
+     */
+    public function transform(mixed $value): mixed
+    {
+        if (!is_string($value)) {
+            return $value;
+        }
+
+        return preg_replace('/[^a-z]+/i', '', $value);
+    }
+}