+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added TransformingValidator to transform a value before validation, a valid result will return the transformed value.
 - Added AlphanumericOnlyTransformer to transform a string to alphanumeric characters with Unicode support.
 - Added AsciiAlphanumericOnlyTransformer to transform a string to alphanumeric characters without Unicode support.
+- Added AsciiLettersOnlyTransformer to transform a string to ASCII letters.
 
 ### Changed